Update GreetingService for 0.2.0 and update docs (#240)
Update GreetingService for 0.2.0 and update docs

### Motivation

The example project needs updating, but it was only possible to do after we've tagged 0.2.0, so here we go.

### Modifications

Updated the example service and updated docs for using the command plugin, which I noticed wasn't documented.

### Result

The example project now uses the latest version. Our tutorials point to it, so it's good for it to track the latest and greatest.

### Test Plan

Built locally.


Reviewed by: gjcairo, glbrntt

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/240