Add support for nullable schemas (#248)
### Motivation

Fixes #82.

More background:

In JSON Schema, there are two fields that affect what we represent as
optionality in Swift.
1. In object schemas, not including a property in the `required` array.
2. In all schemas, marking the schema as nullable. In OpenAPI 3.0, that
used to be done using a dedicated `nullable` field, in OpenAPI 3.1, that
field is removed and instead the same result is achieved by adding the
`null` type in the `type` array. So for a non-nullable string, you'd use
`type: string`, and for a nullable string, you'd use `type: [string,
null]`.

Up until now, we've only supported (1) in how we decide which types are
generated as optional, and that seems to be how most adopters control
nullability. However, (2) is also supported in JSON Schema, so we should
also support it, and as evidenced by the comments in #82, a few folks
have hit the lack of nullability support in Swift OpenAPI Generator.

### Modifications

This PR takes nullability on the schema into account when deciding
whether to generate a type as optional or not.

Since this is a source-breaking change, it's introduced behind the
feature flag `nullableSchemas`, and it'll become default behavior in
0.3.0.

Most of this change is just propagating the feature flag to the right
place.

### Result

Marking a schema as nullable will result in an optional type being
generated.

### Test Plan

Added a snippet test, both with and without the feature flag, which
shows the change.