Compile with -warnings-as-errors in CI (incl. StrictSendability) (#281)
### Motivation

We were not compiling with `-warnings-as-errors` in CI because some of
the reference source code contained deprecation annotations. This meant
that we are not getting CI failures for other warnings that we'd like to
be clean of; specifically strict concurrency. However, in addition to
the file-based reference tests, we now have in-memory snippet-based
reference tests, which contain tests for deprecated OpenAPI schemas and
operations.

### Modifications

- Remove deprecated tests from file-based reference test.
- Clean up warnings under StrictConcurrency.
- Compile with -warnings-as-errors in CI.
- Opt into StrictConcurrency and `-warnings-as-errors` locally with
environment variable.

### Result

- CI will catch sendability warning regressions.
- Developing locally with `SWIFT_OPENAPI_STRICT_CONCURRENCY` will enable
the same behaviour.

### Test Plan

- Added a canary commit to the end of this PR with a sendability issue,
which will be reverted before merging.

---------

Signed-off-by: Si Beaumont <beaumont@apple.com>