Move warning about strict concurrency into PlatformChecks.swift (#289)
### Motivation

In #191 we enabled strict concurrency in CI. We also attempted to
include it conditionally as part of a local development flow. However,
the `#warning` (that strict concurrency was enabled) added to
Package.swift was firing unconditionally because it was, itself, behind
a _runtime_ conditional.

### Modifications

Now if `SWIFT_OPENAPI_STRICT_CONCURRENCY=true` then we will add a
`define` to the `swiftSettings` and we will move the `#warning` to the
existing `PlatformChecks.swift`.

### Result

Opening Xcode with `SWIFT_OPENAPI_STRICT_CONCURRENCY=true` will enable
strict concurrency warnings locally. Opening Xcode without this will no
longer produce a misleading warning.

### Test Plan

Tested locally.

Signed-off-by: Si Beaumont <beaumont@apple.com>