Add url form encoder & decoder (#283)
### Motivation

#182 Currently, the generated interface for request bodies that use
URLEncodedForms accepts a `Data` object. This PR uses the
URIEncoder/Decoder to instead provide an interface using codable types
in the same way that a JSON body would.

### Modifications
Adds a new coding strategy that works with the corresponding [runtime
converter
updates](https://github.com/apple/swift-openapi-runtime/pull/53).

**Note:** This PR does not support encoding collections of primitives
(i.e. an array of string values for a single key). This should be
addressed in a future update, but it raised questions about the correct
use of URIEncoder, so I didn't include it in this PR. Further discussion
of the issue is in the [runtime
PR](https://github.com/apple/swift-openapi-runtime/pull/53).

### Result
Generated interfaces now directly accept URLEncodedForm Codable types
rather than untyped `Data`.

### Test Plan

Updated reference tests to include a request using a form body.

---------

Co-authored-by: benfrearson <ben.frearson@sonin.agency>
Co-authored-by: bfrearson <>
Co-authored-by: Honza Dvorsky <honza@apple.com>