YAML diagnostics collector should have stable output (#295)
YAML diagnostics collector should have stable output

### Motivation

Fixes #294.

Since we're running generation in parallel now, the order of diagnostics depends on the racing of the cores, resulting in unstable output.

### Modifications

Sort the diagnostics before writing them to file, making the output stable again.

### Result

Stable YAML diagnostics file.

### Test Plan

N/A, the order isn't significant, so we're not testing the details of the sorting function, we just care about it being stable.


Reviewed by: dnadoba

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (compatibility test) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/295