Propagate nullability through references (#300)
Propagate nullability through references

### Motivation

Fixes #288.

### Modifications

Refactors how we decide whether to mark a type as optional, and adds the ability to "peek" through references to decide the nullability of the reference schema.

Also added an article describing the overall logic as a starting point, will be expanded with a few more changes coming in this area in future PRs.

### Result

We now correctly mark an object property as optional if it points at an optional schema through a reference. Previously, it was always marked as required, which caused decoding to fail e.g. in the GitHub OpenAPI doc.

### Test Plan

Added and adjusted unit tests.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (compatibility test) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 
     ✖︎ pull request validation (integration test) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/300