Update all examples and integration tests to 0.3.0 (#325)
Update all examples and integration tests to 0.3.0

### Motivation

We're ready to release 0.3.0, so update all examples and integration tests, so that once tagged, the code uses the latest versions.

### Modifications

Updated integration tests and examples, also made a few minor improvements.

### Result

All code refers to 0.3.0.

### Test Plan

Verified manually, but some of these might not pass because 0.3.0 of the generator doesn't exist yet - chicken and egg, but that's understood as part of this process.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.10) - Build finished. 
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (compatibility test) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 
     ✖︎ pull request validation (integration test) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/325