Fix inline response headers (#355)
Fix inline response headers

### Motivation

When generating a response header that points to an inline schema, the generator didn't generate the definition of the inline schema, so the code wouldn't compile.

### Modifications

Make changes to properly compute optionality (drive-by improvement) and generate the inline type, if needed.

### Result

Inline response header schemas now compile.

### Test Plan

Added a snippet test for this.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.10) - Build finished. 
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (compatibility test) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/355