Bump OpenAPIKit to 3.0.0-rc.3 and recognize more base64/binary encoding locations (#357)
Bump OpenAPIKit to 3.0.0-rc.3 and recognize more base64/binary encoding locations

### Motivation

In OpenAPI 3.1.0, the `format: ...` part of schemas is deprecated, and instead `contentEncoding: ...` should be used. Before this PR, we'd recognize the encodings defined in `format` but not `contentEncoding`, this PR fixes that and finds them in both locations. This required a bump of OpenAPIKit, which added support for `contentEncoding` in `3.0.0-rc.3`.

### Modifications

Change the type matching logic to first look at `contentEncoding`, if no special value is specified there, look at `format`.

### Result

`contentEncoding: base64` and `contentEncoding: binary` will now be correctly represented as their runtime types instead of a plain string.

### Test Plan

Adapted the unit tests.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.10) - Build finished. 
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (compatibility test) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/357