Bump to OpenAPIKit 3.0.0 (#362)
Bump to OpenAPIKit 3.0.0

### Motivation

OpenAPIKit released 3.0.0, so let's move to it.

### Modifications

They (correctly) removed the `format: byte` and `format: binary` fields in the OpenAPI 3.1 mode, instead `contentEncoding: base64` and `contentEncoding: binary` should be used.

For docs that load as 3.0 and get converted to 3.1 in memory, this translation happens automatically in OpenAPIKit, so we just need to be able to handle 3.1 documents here.

### Result

Using the stable v3 version of OpenAPIKit, which will allow Swift OpenAPI Generator to coexist with other projects that depend on OpenAPIKit v3 (previously, locking to an exact version of OpenAPIKit meant more difficult version resolution for adopters.)

### Test Plan

Adapted tests, all pass now.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.10) - Build finished. 
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (compatibility test) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/362