Fix a bug in path parameter ordering (#380)
### Motivation

Fixes #378.

Basically, if the path parameter order was different between the URL
template `/a/{a}/b/{b}` and the list of path parameters, for example if
`b` came before `a` there, we used the parameter order, not the URL
template order, which was _incorrect_. I suspect this bug has been
hiding for this long is because most documents use the same order in
both.

### Modifications

Change the algorithm that composes the URL template and the list of
parameters to respect the URL template's order.

### Result

The URL template order is now respected, regardless of the path
parameter order.

### Test Plan

Added a snippet test that deliberately uses a different order between
the URL template and the path paramater list. It failed before this fix,
passes with it.