Undeprecate tests that used to test deprecated code (#395)
Undeprecate tests that used to test deprecated code

### Motivation

Two tests used to exercise operations in the Petstore test project that were deprecated, so to avoid warnings we deprecated the tests themselves too. We removed the deprecation of those operations a while ago, and forgot to undeprecate the tests.

### Modifications

Undeprecate the tests.

### Result

Removed an unnecessary deprecation of two tests.

### Test Plan

All tests pass.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.10) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (compatibility test) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 
     ✖︎ pull request validation (5.8) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/395