Add visionOS to Package.swift to emit useful compilation errors (#398)
### Motivation

We have provision for emitting a useful compilation error when trying to
link the _generator_ to an application on darwin platforms, e.g. iOS,
tvOS, watchOS. To support this, we add those platforms to the
Package.swift so that the code at least tries to compile and then we can
emit the `#error`. This list does not currently include visionOS so
users will not get the same useful compilation error.

### Modifications

Add visionOS to the platforms in Package.swift. While there, also add
this to the platforms in the Package.swift for the client used in the
tutorial.

### Result

When compiling for visionOS and trying to link the generator, a useful
compiler error will be emitted.

### Test Plan

Manual.