Bump to 1.0.0-alpha.1 (#400)
Bump to 1.0.0-alpha.1

### Motivation

Prep for 1.0.0-alpha.1.

### Modifications

Bumped documentation and upstream SOAR dependencies to 1.0.0-alpha.1.

Note that because the generator itself is not yet released, the integration test and examples will fail to build, but that's the unfortunate nature of having examples projects in the same repo as the project itself. But it's important that once this repo is tagged, that the examples/integration tests included use the right tags, so I don't think there's another way to stage this in.

### Result

Ready to tag 1.0.0-alpha.1.

### Test Plan

Temporarily modified the generator dependency in examples/integration tests to point to `main` and verified all works. Then changed it back.

All other tests pass.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.10) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (compatibility test) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 
     ✖︎ pull request validation (integration test) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/400