scripts: Use shared --cache-path and --scratch-path when building examples (#436)
### Motivation

In a small attempt to make the example pipeline run faster, we can share
the cache and scratch paths. This might shave a few minutes off the CI
run, until we have time to investigate it further. FWICT, it seems that
there are still a number of builds of dependencies (e.g. NIO) flying
past.

### Modifications

Use shared --cache-path and --scratch-path when building examples.

It's only shared for the single run of the script.

### Result

Slightly faster CI.

### Test Plan

CI.