scripts: Use git-archive to copy example to avoid ignored files (#434)
### Motivation

The `test-examples.sh` script was using `cp -R` to copy the example path
to a temporary directory before building. This brought along files that
weren't checked in, and files that are in `.gitignore`. This resulted in
a build failure for me when running locally because I had been iterating
within that directory using `swift build` and when I ran the script, it
copied over the `.build/` directory. But, in principle, we should only
take the indexed files since these are what will be in the repo, and,
therefore, what we expect users to copy and be able to build.

### Modifications

Use `git archive` to copy the contents of the specific example to the
temporary directory.

### Result

Only files tracked by Git will be taken at the time of running
`test-examples.sh`.

### Test Plan

CI should still pass.

Signed-off-by: Si Beaumont <beaumont@apple.com>