Reword examples parent README with disclaimer (#445)
### Motivation

As we approach 1.0, we want more examples of how to use this package and
integrate with other packages in the ecosystem. However, these are
deliberately meant to be illustrative, and are often deliberately not
fully-featured. The README currently suggests using these as starter
projects, which implies something stronger.

### Modifications

Reword examples parent README with disclaimer
 
### Result

Docs clearer.

### Test Plan

None.