Update NOTICE.txt with projects used in examples (#463)
### Motivation

We've recently added a bunch of example projects that show how to
integrate with other packages and projects in the Swift and OpenAPI
ecosystems. We should make sure that these projects and their associated
licenses are mentioned in `NOTICE.txt`.

### Modifications

Update NOTICE.txt with projects used in examples.

### Result

Updated NOTICE.txt with projects used in examples.

### Test Plan

None.