Proposal documentation cleanup and consistency (#469)
### Motivation

We want to go into 1.0 with the proposal documentation up to date,
w.r.t. the
version in which the proposal was implemented, and linking to the forums
post
where they were reviewed. They also have diverged a bit in
style.

### Modifications

Update the proposal docs for consistent style. I referred again to the
style
used in the Swift Evolution repo.

### Result

Proposal docs consistent and up to date.

### Test Plan

None.