Add Github issue template forms (#477)
### Motivation

Github recently added support for Issue forms, which can help guide the
creation of issues for adopters and developers to provide the critical
information needed to diagnose issues. These have recently been adopted
by the Swift project too, from which our templates take much
inspiration.

When using issue forms (and templates in general) it also provides the
filer with a choice of what kind of issue it is, and can automatically
add appropriate labels based on kind and apply the `status/triage`
label.

Finally the page that displays the choice of issue templates can be
further extended to provide links to other useful external resources.

Note that it is still possible to create an issue from a blank slate if
none of the options suit the adopter.

Check out the "New issue" page on apple/swift to see all these features
combined and in action[^1].

[^1]: https://github.com/apple/swift/issues/new/choose

### Modifications

- Add template for reporting a bug:
  - Disclaimer that this is not for Xcode issues. 
  - Required fields for versions, reproduction steps, expected behavior.
  - Optional fields for additional information.
  - Applies labels: `kind/bug`, `status/triage`.
- Add template for asking a question:
  - Disclaimer that this is not for Xcode support. 
  - Links to FAQ.
  - Applies labels: `kind/support`, `status/triage`.
- Add template for requesting a feature:
  - Explains the proposal process.
  - Required fields for cut-down proposal input.
  - Applies labels: `kind/feature`, `status/triage`.
- Add template for tracking a task:
- Meant for maintainers and for breaking out issues from PRs or quotes
from bugs.
  - Applies labels: `kind/enhancement`, `status/triage`.
- Added additional resources to chooser:
  - Documentation
  - Examples
  - FAQ
  - Supported OpenAPI features
  - Proposal process
  - Swift Open Source Slack channel
  - Feedback Assistant for Xcode and URLSession bugs
- Kept the blank issues enabled.

### Result

More guided process when filing issues and surfaces other relevant
resources.

### Test Plan

None.