[Docs] Refresh tutorials (#478)
### Motivation

Our tutorials were originally written for SOAR 0.1.0, so we need to
update them to reflect the best practices as of SOAR 1.0 instead.

### Modifications

- Adapted to the new Examples structure (most importantly, point to
cloning the Vapor hello world project when writing the client, instead
of the GreetingService.)
- Added a mention of the shorthand APIs as an alternative to the
exhaustive handling. (Still show exhaustive handling first though, as we
want to make sure adopters consider that first and even prefer it.)
- More minor polish and updates.

### Result

Tutorials ready for 1.0 (sans versions, those are changed in #472)

### Test Plan

Previewed locally, actually went through all the tutorials and verified
they all work.