[Examples] Stop linking transitive dependencies in tests (#487)
### Motivation

In Swift 5.9.2, a bug is fixed that forced us to previously explicitly
link transitive dependencies in tests. So let's remove it, and also bump
the swift-tools-version of the affected projects to 5.9.2 to make it
explicit.

### Modifications

Removed the explicit linkage, bumped the swift-tools-version.

### Result

Examples show the best practices using the latest tools.

### Test Plan

Ran locally using:

```
docker run --rm -it -v $PWD/Examples/hello-world-hummingbird-server-example:/app swift:5.9 swift test --package-path /app
```

Failed with 5.9.1. Passed with 5.9.2 (and 5.9, which is now an alias for
5.9.2).