Generate arrays with an optional element type for arrays with nullable items (#492)
### Motivation

Array items can be nullable. When this is the case, the element type of
the generated array should be optional. Today it is not. For example,
consider the following OpenAPI snippet:

```yaml
  StringArrayNullableItems:
    type: array
    items:
      type: [string, null]
```

This currently generates a type alias for `[String]` but it should be
`[String?]`

### Modifications

Update the translator to take into account `arrayContext.items.nullable`
when generating the types for array values.

### Result

Arrays with nullable items 

### Test Plan

- Snippet test for standalone schema.
- Snippet test for use within an object.
- A test that shows lossless conversion to JSON with Foundation.

### Related issues

- Fixes #444.